CBG-3851 move AssertLogContains tests to run without race detector
AssertLogContains swaps the global logger which can trigger a data race
if other things are writing to the logger (e.g. walrus or gocb
background tasks)

The benefit of running race on these tests would be marginal.
